requirements.txt
aiogram==3.0.0
python-dotenv==0.19.0
psycopg2-binary==2.9.3
aiohttp==3.8.1